requirements.txt
accelerate==1.11.0
annotated-types==0.7.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.1
et_xmlfile==2.0.0
filelock==3.19.1
fsspec==2025.9.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
joblib==1.5.2
lightning-utilities==0.15.2
MarkupSafe==2.1.5
mpmath==1.3.0
narwhals==2.11.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
open_clip_torch==3.2.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
peft==0.17.1
pillow==11.3.0
platformdirs==4.5.0
plotly==6.4.0
protobuf==6.33.1
psutil==7.1.2
pycocotools==2.0.10
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
randaugment==1.0.2
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers>=2.7.0
sentry-sdk==2.46.0
six==1.17.0
smmap==5.0.2
sympy==1.13.1
threadpoolctl==3.6.0
timm==1.0.21
tokenizers==0.22.1
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchmetrics==1.8.2
torchvision==0.21.0+cu118
tqdm==4.67.1
transformers==4.57.1
triton==3.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.23.0
wcwidth==0.2.14
